+++ src/actions/actionTypes.js
@@ -1,6 +1,2 @@
-export const TOAST = 'TOAST';
-export const TOASTED = 'TOASTED';
 export const LOADING = 'LOADING';
 export const GET_TOPICS_SUCCESS = 'GET_TOPICS_SUCCESS';
-export const CANCEL_TARGET = 'CANCEL_TARGET';
-export const UPDATE_TARGETS = 'UPDATE_TARGETS';
