+++ src/actions/sessionActions.js
@@ -3,18 +3,6 @@ import { sessionService } from 'redux-react-session';
 import { toast } from 'react-toastify';
 
 import sessionApi from '../api/sessionApi';
-import * as types from './actionTypes';
-import history from '../utils/history';
-import routes from '../constants/routesPaths';
-
-const getTopicsSucess = ({ topics }) => ({
-  type: types.GET_TOPICS_SUCCESS,
-  topics
-});
-
-const loading = () => ({
-  type: types.LOADING
-});
 
 export const login = user =>
   () =>
@@ -45,24 +33,3 @@ export const logout = () =>
     }).catch((err) => {
       throw (err);
     });
-
-export const editUser = user =>
-  () =>
-    sessionApi.updateUser(user)
-      .then(({ user }) => {
-        toast('Account information successfully updated!');
-        history.push(routes.index);
-        sessionService.saveUser(user);
-      })
-      .catch((err) => {
-        throw new SubmissionError({
-          _error: err.errors
-        });
-      });
-
-export const getTopics = () => (dispatch) => {
-  dispatch(loading());
-  sessionApi.getTopics().then((topics) => {
-    dispatch(getTopicsSucess(topics));
-  }).catch(() => toast.error('Topics could not be retrieved'));
-};
