+++ src/actions/userActions.js
@@ -1,7 +1,10 @@
 import { SubmissionError } from 'redux-form/immutable';
 import { sessionService } from 'redux-react-session';
+import { toast } from 'react-toastify';
 
 import sessionApi from '../api/sessionApi';
+import history from '../utils/history';
+import routes from '../constants/routesPaths';
 
 export const signUp = user =>
   () =>
@@ -12,3 +15,17 @@ export const signUp = user =>
         _error: err.errors
       });
     });
+
+export const editUser = user =>
+  () =>
+    sessionApi.updateUser(user)
+      .then(({ user }) => {
+        toast('Account information successfully updated!');
+        history.push(routes.index);
+        sessionService.saveUser(user);
+      })
+      .catch((err) => {
+        throw new SubmissionError({
+          _error: err.errors
+        });
+      });
