+++ src/api/sessionApi.js
@@ -4,11 +4,9 @@ class Session {
   static login(user) {
     return api.post('/users/sign_in', user);
   }
-
   static loginFacebook(user) {
     return api.post('/users/facebook', user);
   }
-
   static logout() {
     return api.delete('/users/sign_out');
   }
@@ -18,15 +16,6 @@ class Session {
   static updateUser(user) {
     const { id } = user;
     return api.put(`users/${id}`, { user });
-  }
-  static getTopics() {
-    return api.get('topics');
-  }
-  static createTarget(target) {
-    return api.post('targets', { target });
-  }
-  static getTargets() {
-    return api.get('targets', { page: 1 });
   }
 }
