+++ src/api/topicApi.js
@@ -1,0 +1,9 @@
+import api from './apiService';
+
+class Topic {
+  static getTopics() {
+    return api.get('topics');
+  }
+}
+
+export default Topic;
