--- src/utils/test.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import HomePage from '../containers/HomePage';
-
-const home = props => (<HomePage {...props} />);
-
-export default home;
